requirements.txt
langgraph
langchain
openai
streamlit
mysql-connector-python
langchain
pydantic
python-dotenv
ollama
chromadb
